front: Fix crash when rom extraction fails

+++ src/paulscode/android/mupen64plusae/GalleryActivity.java
@@ -1042,8 +1042,8 @@ public class GalleryActivity extends AppCompatActivity implements GameSidebarAct
                         final String computedMd5 = ComputeMd5Task.computeMd5( tempRomPath );
                         lbFound = computedMd5 != null && computedMd5.equals(md5);
 
-                        //only deleye the file if we extracted our selves
-                        if(!lbFound && !fileExisted)
+                        //only delete the file if we extracted our selves
+                        if(!lbFound && !fileExisted && tempRomPath != null)
                         {
                             tempRomPath.delete();
                         }
